requirements.txt
langchain==0.1.14
langgraph==0.0.31
tavily-python==0.3.3
BeautifulSoup4==4.12.3
openai==1.16.1
flask==3.0.2
json5==0.9.24
Flask-CORS==4.0.0
langchain_openai==0.1.1
python-dotenv==1.0.1
ipython==8.23.0
grandalf==0.8